[MWX Checkpoint] Before editing d:\Tensor-Engine\src\nn\transformer.rs (2026-06-13T19-27-14-365Z)

+++ src/nn/transformer.rs
@@ -740,7 +740,7 @@ impl MultiHeadAttention {
                 if causal {
                     // mask shape: (b*num_heads, q_seq, kv_seq)
                     let mut mask_arr = ndarray::ArrayD::<f32>::zeros(ndarray::IxDyn(
-                        &[b * self.num_heads, q_seq, kv_seq][..],
+                        &[b as isize * self.num_heads as isize, q_seq as isize, kv_seq as isize][..],
                     ));
                     let new_start = kv_seq.saturating_sub(q_seq);
                     for i in 0..(b * self.num_heads) {
